fix: moved chiPT stuff to separate folder
